// Assets-Uncapped/front-portal-develop/src/services/api/organisation-users/models/ChangeUserPhoneRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * api-organisation-users
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: @project.version@
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface ChangeUserPhoneRequest
 */
export interface ChangeUserPhoneRequest {
    /**
     * 
     * @type {string}
     * @memberof ChangeUserPhoneRequest
     */
    phone?: string;
    /**
     * 
     * @type {boolean}
     * @memberof ChangeUserPhoneRequest
     */
    smsConsentGranted?: boolean;
}

/**
 * Check if a given object implements the ChangeUserPhoneRequest interface.
 */
export function instanceOfChangeUserPhoneRequest(value: object): value is ChangeUserPhoneRequest {
    return true;
}

export function ChangeUserPhoneRequestFromJSON(json: any): ChangeUserPhoneRequest {
    return ChangeUserPhoneRequestFromJSONTyped(json, false);
}

export function ChangeUserPhoneRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): ChangeUserPhoneRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'phone': json['phone'] == null ? undefined : json['phone'],
        'smsConsentGranted': json['smsConsentGranted'] == null ? undefined : json['smsConsentGranted'],
    };
}

export function ChangeUserPhoneRequestToJSON(json: any): ChangeUserPhoneRequest {
    return ChangeUserPhoneRequestToJSONTyped(json, false);
}

export function ChangeUserPhoneRequestToJSONTyped(value?: ChangeUserPhoneRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'phone': value['phone'],
        'smsConsentGranted': value['smsConsentGranted'],
    };
}